Rename minimax file to alphabeta
